# openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
  /api/v1/sales/targets/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateTarget
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesTargetRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesTargetDto'
    delete:
      tags:
      - sales-controller
      operationId: deleteTarget
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  /api/v1/sales/promotions/{id}:
    put:
      tags:
      - sales-controller
      operationId: updatePromotion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePromotionDto'
    delete:
      tags:
      - sales-controller
      operationId: deletePromotion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  /api/v1/sales/orders/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesOrderRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesOrderDto'
    delete:
      tags:
      - sales-controller
      operationId: deleteOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  /api/v1/sales/credit-requests/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateCreditRequest
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseCreditRequestDto'
...
  /api/v1/dealers/{id}:
    put:
      tags:
      - sales-controller
      operationId: updateDealer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseDealerDto'
    delete:
      tags:
      - sales-controller
      operationId: deleteDealer
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
  /api/v1/sales/targets:
    get:
      tags:
      - sales-controller
      operationId: targets
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListSalesTargetDto'
    post:
      tags:
      - sales-controller
      operationId: createTarget
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesTargetRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesTargetDto'
  /api/v1/sales/promotions:
    get:
      tags:
      - sales-controller
      operationId: promotions
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListPromotionDto'
    post:
      tags:
      - sales-controller
      operationId: createPromotion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponsePromotionDto'
  /api/v1/sales/orders:
    get:
      tags:
      - sales-controller
      operationId: orders
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListSalesOrderDto'
    post:
      tags:
      - sales-controller
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesOrderRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesOrderDto'
  /api/v1/sales/orders/{id}/confirm:
    post:
      tags:
      - sales-controller
      operationId: confirmOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesOrderDto'
  /api/v1/sales/orders/{id}/cancel:
    post:
      tags:
      - sales-controller
      operationId: cancelOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesOrderDto'
  /api/v1/sales/credit-requests:
    get:
      tags:
      - sales-controller
      operationId: creditRequests
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListCreditRequestDto'
    post:
      tags:
      - sales-controller
      operationId: createCreditRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseCreditRequestDto'
...
  /api/v1/dealers:
    get:
      tags:
      - sales-controller
      operationId: dealers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListDealerDto'
    post:
      tags:
      - sales-controller
      operationId: createDealer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealerRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseDealerDto'
...
  /api/v1/sales/orders/{id}/status:
    patch:
      tags:
      - sales-controller
      operationId: updateStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseSalesOrderDto'
...
  /api/v1/orchestrator/dashboard/finance:
    get:
      tags:
      - dashboard-controller
      operationId: financeDashboard
      parameters:
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/dashboard/factory:
    get:
      tags:
      - dashboard-controller
      operationId: factoryDashboard
      parameters:
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/v1/orchestrator/dashboard/admin:
    get:
      tags:
      - dashboard-controller
      operationId: adminDashboard
      parameters:
      - name: X-Company-Id
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
  /api/v1/invoices/{id}:
    get:
      tags:
      - invoice-controller
      operationId: getInvoice
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseInvoiceDto'
  /api/v1/invoices/dealers/{dealerId}:
    get:
      tags:
      - invoice-controller
      operationId: dealerInvoices
      parameters:
      - name: dealerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListInvoiceDto'
...
components:
  schemas:
    SalesTargetRequest:
      required:
      - name
      - periodEnd
      - periodStart
      - targetAmount
      type: object
      properties:
        name:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        targetAmount:
          type: number
        achievedAmount:
          type: number
        assignee:
          type: string
    ApiResponseSalesTargetDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/SalesTargetDto'
        timestamp:
          type: string
          format: date-time
    SalesTargetDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        targetAmount:
          type: number
        achievedAmount:
          type: number
        assignee:
          type: string
    PromotionRequest:
      required:
      - discountType
      - discountValue
      - endDate
      - name
      - startDate
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
    ApiResponsePromotionDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/PromotionDto'
        timestamp:
          type: string
          format: date-time
    PromotionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
    SalesOrderItemRequest:
      required:
      - productCode
      - quantity
      - unitPrice
      type: object
      properties:
        productCode:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
    SalesOrderRequest:
      required:
      - totalAmount
      type: object
      properties:
        dealerId:
          type: integer
          format: int64
        totalAmount:
          type: number
        currency:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesOrderItemRequest'
    ApiResponseSalesOrderDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/SalesOrderDto'
        timestamp:
          type: string
          format: date-time
    SalesOrderDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        orderNumber:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
        dealerName:
          type: string
        traceId:
          type: string
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesOrderItemDto'
    SalesOrderItemDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productCode:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
    CreditRequestRequest:
      required:
      - amountRequested
      type: object
      properties:
        dealerId:
          type: integer
          format: int64
        amountRequested:
          type: number
        reason:
          type: string
        status:
          type: string
    ApiResponseCreditRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/CreditRequestDto'
        timestamp:
          type: string
          format: date-time
    CreditRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        dealerName:
          type: string
        amountRequested:
          type: number
        status:
          type: string
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
...
    DealerRequest:
      required:
      - code
      - creditLimit
      - name
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        email:
          type: string
        phone:
          type: string
        creditLimit:
          type: number
    ApiResponseDealerDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/DealerDto'
        timestamp:
          type: string
          format: date-time
    DealerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        creditLimit:
          type: number
        outstandingBalance:
          type: number
...
    CancelRequest:
      type: object
      properties:
        reason:
          type: string
...
    StatusRequest:
      type: object
      properties:
        status:
          type: string
    ApiResponseListSalesTargetDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/SalesTargetDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListPromotionDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/PromotionDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListSalesOrderDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/SalesOrderDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListCreditRequestDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/CreditRequestDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListInvoiceDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceDto'
        timestamp:
          type: string
          format: date-time
    InvoiceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        status:
          type: string
        subtotal:
          type: number
        taxTotal:
          type: number
        totalAmount:
          type: number
        currency:
          type: string
        issueDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        dealerId:
          type: integer
          format: int64
        dealerName:
          type: string
        salesOrderId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineDto'
    InvoiceLineDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productCode:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        taxRate:
          type: number
        lineTotal:
          type: number
    ApiResponseInvoiceDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/InvoiceDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListDealerDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/DealerDto'
        timestamp:
          type: string
          format: date-time
...